FR-1453 re-organize statemachine functions by placing non-interface functions in helper
